算力分析/NNparser-master/requirements.txt
pytorch==1.5.0
tensorflow==2.1.0
pandas
numpy
matplotlib
openpyxl
scikit-learn
scikit-image
graphviz
python-graphviz
pydot
torchvision
